backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
cryptography
flask-jwt-extended
python-dotenv
bcrypt
gunicorn
pymysql
